Move tests to test_bits
